Workout_Tracker/requirements.txt
uvicorn
fastapi
pydantic
starlette
pymysql
sqlalchemy
alembic
passlib
bcrypt
python-multipart
python-jose[cryptography]
dotenv